# backend/tests/fixtures/openapi/breaking.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LoginRequest": {
        "description": "用户登录请求",
        "properties": {
          "password": {
            "description": "密码",
            "maxLength": 100,
            "title": "Password",
            "type": "string"
          },
          "username": {
            "description": "用户名",
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username",
          "password"
        ],
        "title": "LoginRequest",
        "type": "object"
      },
...
      "RegisterRequest": {
        "description": "用户注册请求",
        "properties": {
          "email": {
            "maxLength": 100,
            "minLength": 3,
            "pattern": "^[^@\\s]+@[^@\\s]+\\.[^@\\s]+$",
            "title": "Email",
            "type": "string"
          },
          "password": {
            "description": "密码",
            "maxLength": 100,
            "minLength": 8,
            "title": "Password",
            "type": "string"
          },
          "username": {
            "description": "用户名",
            "maxLength": 50,
            "minLength": 3,
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username",
          "email",
          "password"
        ],
        "title": "RegisterRequest",
        "type": "object"
      },
...
      "TokenResponse": {
        "description": "登录成功响应",
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "token_type": {
            "default": "bearer",
            "title": "Token Type",
            "type": "string"
          },
          "user_id": {
            "title": "User Id",
            "type": "integer"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "access_token",
          "user_id",
          "username"
        ],
        "title": "TokenResponse",
        "type": "object"
      },
      "UserResponse": {
        "description": "用户信息响应",
        "properties": {
          "email": {
            "title": "Email",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "id",
          "username",
          "email",
          "is_active"
        ],
        "title": "UserResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "AI 辅助小说创作与理解平台",
    "title": "NovelMind API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/auth/login": {
      "post": {
        "description": "用户登录。\n\n验证用户名和密码，成功后返回 JWT Access Token。",
        "operationId": "login_api_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login",
        "tags": [
          "认证"
        ]
      }
    },
...
    "/api/auth/me": {
      "get": {
        "description": "获取当前登录用户信息",
        "operationId": "get_me_api_auth_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Me",
        "tags": [
          "认证"
        ]
      }
    },
    "/api/auth/register": {
      "post": {
        "description": "用户注册。\n\n检查用户名和邮箱是否已存在，对密码进行 bcrypt 哈希后存入数据库。",
        "operationId": "register_api_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register",
        "tags": [
          "认证"
        ]
      }
    },
    "/api/characters/{novel_id}": {
      "get": {
        "description": "获取小说人物列表",
        "operationId": "get_characters_api_characters__novel_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "novel_id",
            "required": true,
            "schema": {
              "title": "Novel Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Characters",
        "tags": [
          "人物关系"
        ]
      }
    },
    "/api/characters/{novel_id}/extract": {
      "post": {
        "description": "AI 自动识别人物及关系（NER + LLM 混合抽取）",
        "operationId": "extract_characters_api_characters__novel_id__extract_post",
        "parameters": [
          {
            "in": "path",
            "name": "novel_id",
            "required": true,
            "schema": {
              "title": "Novel Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Extract Characters",
        "tags": [
          "人物关系"
        ]
      }
    },
    "/api/characters/{novel_id}/relations": {
      "get": {
        "description": "获取人物关系网络（用于 AntV G6 图谱可视化）",
        "operationId": "get_relations_api_characters__novel_id__relations_get",
        "parameters": [
          {
            "in": "path",
            "name": "novel_id",
            "required": true,
            "schema": {
              "title": "Novel Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Relations",
        "tags": [
          "人物关系"
        ]
      }
    },
...
    "/api/fanfiction": {
      "post": {
        "description": "创建同人文（指定原作、标题、续写提示）",
        "operationId": "create_fanfiction_api_fanfiction_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Data",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Fanfiction",
        "tags": [
          "同人文"
        ]
      }
    },
    "/api/fanfiction/{fanfiction_id}/continue": {
      "post": {
        "description": "AI 续写（流式输出，基于 RAG 上下文注入）",
        "operationId": "continue_writing_api_fanfiction__fanfiction_id__continue_post",
        "parameters": [
          {
            "in": "path",
            "name": "fanfiction_id",
            "required": true,
            "schema": {
              "title": "Fanfiction Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Prompt",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Continue Writing",
        "tags": [
          "同人文"
        ]
      }
    },
    "/api/fanfiction/{novel_id}": {
      "get": {
        "description": "获取小说的同人文列表",
        "operationId": "list_fanfictions_api_fanfiction__novel_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "novel_id",
            "required": true,
            "schema": {
              "title": "Novel Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Fanfictions",
        "tags": [
          "同人文"
        ]
      }
    },
...
  }
}